renamed to brute_force_ancestors.rs
